+++ py/ate/README.md
@@ -3,6 +3,8 @@
 Arcyd tester - command-line utility to exercise arcyd.
 * `atet_arcyd_instance.py` -
 Tools to start and manage an Arcyd instance.
+* `atet_sharedrepo.py` -
+Simulates a central repository where multiple workers can collaborate.
 * `atet_worker.py` -
 Utilities to simulate a person working with Phabricator, Git and Arcyd.
